store_count stores long long int values
As now count_t is adjusted to long long int, we can safely store long
long int instead of just int, which actually fixes some overflow with
uni_probs[0].

+++ src/counts.c
@@ -28,7 +28,7 @@ void store_count(flag four_byte_counts,
 		 unsigned short *short_counts,
 		 long long int *long_counts,
 		 int position,
-		 int count) {
+		 long long int count) {
 
   if (four_byte_counts) {
     long_counts[position] = count;
